Fixing a NPE that occurs in Admin
Fixing a NPE that occurs in Admin when navigating to the My Changes and Workflow Audit sections, after making changes to a Content Item. One example of when the NPE occurs is when you edit a content item and try to approve it through the sandboxing workflow. The class is `org.broadleafcommerce.cms.structure.domain.StructuredContent` which returns no sections, and the NPE is thrown on in the for loop because `sections` is null